+++ backend/reitera-backend/src/main/java/com/helderruiz/reitera_backend/modules/user/dto/AuthResponseDTO.java
@@ -1,7 +1,8 @@
 package com.helderruiz.reitera_backend.modules.user.dto;
 
 public record AuthResponseDTO(
-        String token,
+        String accessToken,
+        String refreshToken,
         String message
 ) {
 }